refactor(session): extract SessionPersistenceService and SessionEncryptionService (#4251 step 1+2)
- Move load/save/serialize/flush/backup from SessionManager to SessionPersistenceService
- Move encrypt/decrypt/setKey from SessionManager to SessionEncryptionService
- SessionManager now delegates to both services via constructor injection
- Removed 177 lines from session.ts (1648 → ~1471)
- 3 new tests for persistence, 468 existing tests still pass
- No behavioral changes — pure refactor
